feat: imagePickerManager 생성
